Extract repeated spec functions for testing summary lists into a helper
This reduces duplication and makes it easier to reuse them in other specs
